chore(index filter): move out of bb bus folder
SVA-382
