+++ src/components/index.js
@@ -34,6 +34,7 @@ export * from './Image';
 export * from './ImageSection';
 export * from './ImagesCarousel';
 export * from './ImagesCarouselItem';
+export * from './IndexFilterWrapperAndList';
 export * from './Label';
 export * from './Link';
 export * from './ListComponent';
